// src/app/calculator/api/model/inboundDto.ts
/**
 * Azure Functions OpenAPI Extension
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface InboundDto { 
    powierzchniaZlewni?: number;
    glebokoscWolna?: number;
    glebokoscOgroduDeszczowego?: number;
    qSplywuDla15lsha?: number;
    qSplywuDla130lsha?: number;
    qSplywuDla300lsha?: number;
}