// src/app/calculator/api/model/outboundDto.ts
/**
 * Azure Functions OpenAPI Extension
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface OutboundDto { 
    minPowierzchnia?: number;
    objetoscOpaduPierwszaFala?: number;
    objetoscOpadu130lsha?: number;
    objetoscOpadu300lsha?: number;
    objetoscOgroduDeszczowego?: number;
    statusPierwszaFala?: boolean;
    statusOpad130lsha?: boolean;
    statusOpad300lsha?: boolean;
    calculationStatus?: OutboundDto.CalculationStatusEnum;
}
export namespace OutboundDto {
    export type CalculationStatusEnum = 1 | 2 | 3;
    export const CalculationStatusEnum = {
        NUMBER_1: 1 as CalculationStatusEnum,
        NUMBER_2: 2 as CalculationStatusEnum,
        NUMBER_3: 3 as CalculationStatusEnum
    };
}